test(IDENT-3086): Add tests for multi-tenancy managers
